requirements.txt
python-telegram-bot==20.7
requests==2.31.0
beautifulsoup4==4.12.3
selenium==4.16.0
python-dotenv==1.0.0
webdriver-manager==4.0.1